removed builder functions from tests
